fix(db): move db connect to storage package
