+++ internal/storage/db/postgresql/options.go
@@ -1,4 +1,4 @@
-package model
+package postgresql
 
 import "time"
