+++ internal/storage/db/postgresql/postgresql.go
@@ -1,11 +1,10 @@
-package model
+package postgresql
 
 import (
 	"fmt"
 	"log/slog"
 	"time"
 
-	_ "github.com/jackc/pgx/stdlib" // use as driver for sqlx
 	"github.com/jmoiron/sqlx"
 
 	"github.com/Employee-s-file-cabinet/backend/pkg/e"
@@ -30,8 +29,8 @@ type DB struct {
 	connTimeout     time.Duration
 }
 
-// New создаёт объект DB с заданными параметрами и подключается к БД.
-func New(dsn string, opts ...Option) (*DB, error) {
+// NewDB создаёт объект DB с заданными параметрами и подключается к БД.
+func NewDB(dsn string, opts ...Option) (*DB, error) {
 	db := &DB{
 		maxConnIdleTime: defaultMaxConnIdleTime,
 		maxConnLifetime: defaultMaxConnLifetime,
@@ -68,7 +67,7 @@ func New(dsn string, opts ...Option) (*DB, error) {
 	return db, nil
 }
 
-// Shutdown дожидается завершения запросов и закрывает все открытые соединения.
-func (db *DB) Shutdown() error {
+// CloseConnection дожидается завершения запросов и закрывает все открытые соединения.
+func (db *DB) CloseConnection() error {
 	return db.Close()
 }
